# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
  /api/v1/pos_closing_entries:
    get:
      tags:
        - PosClosingEntries
      summary: List PosClosingEntries
      description: Get a paginated list of PosClosingEntries with optional filtering and sorting
      operationId: listPosClosingEntries
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PosClosingEntryList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - PosClosingEntries
      summary: Create PosClosingEntry
      description: Create a new PosClosingEntry
      operationId: createPosClosingEntry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePosClosingEntryRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PosClosingEntry'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/pos_cash_movements:
    get:
      tags:
        - PosCashMovements
      summary: List PosCashMovements
      description: Get a paginated list of PosCashMovements with optional filtering and sorting
      operationId: listPosCashMovements
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PosCashMovementList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - PosCashMovements
      summary: Create PosCashMovement
      description: Create a new PosCashMovement
      operationId: createPosCashMovement
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePosCashMovementRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PosCashMovement'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
    PosClosingEntry:
      type: object
      required:
        - id
        - company_id
        - pos_profile_id
        - opening_entry_id
        - closed_at
        - cashier_party_id
        - grand_total
        - invoice_count
        - difference_total
        - status
        - metadata
      properties:
        id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        pos_profile_id:
          type: string
          format: uuid
        opening_entry_id:
          type: string
          format: uuid
        closed_at:
          type: string
          format: date-time
        cashier_party_id:
          type: string
          format: uuid
        totals_by_method:
          nullable: true
          type: object
        grand_total:
          type: string
        invoice_count:
          type: integer
          format: int32
        difference_total:
          type: string
        status:
          type: string
        metadata:
          type: object
...
    CreatePosClosingEntryRequest:
      type: object
      required:
        - company_id
        - pos_profile_id
        - opening_entry_id
        - closed_at
        - cashier_party_id
        - grand_total
        - invoice_count
        - difference_total
        - status
        - metadata
      properties:
        company_id:
          type: string
          format: uuid
        pos_profile_id:
          type: string
          format: uuid
        opening_entry_id:
          type: string
          format: uuid
        closed_at:
          type: string
          format: date-time
        cashier_party_id:
          type: string
          format: uuid
        totals_by_method:
          nullable: true
          type: object
        grand_total:
          type: string
        invoice_count:
          type: integer
          format: int32
        difference_total:
          type: string
        status:
          type: string
        metadata:
          type: object
...
    PosClosingEntryList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/PosClosingEntry'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    PosCashMovement:
      type: object
      required:
        - id
        - company_id
        - pos_profile_id
        - opening_entry_id
        - cashier_party_id
        - movement_type
        - amount
        - moved_at
        - metadata
      properties:
        id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        pos_profile_id:
          type: string
          format: uuid
        opening_entry_id:
          type: string
          format: uuid
        cashier_party_id:
          type: string
          format: uuid
        movement_type:
          type: string
        amount:
          type: string
        reason:
          nullable: true
          type: string
          maxLength: 300
        moved_at:
          type: string
          format: date-time
        metadata:
          type: object
...
    CreatePosCashMovementRequest:
      type: object
      required:
        - company_id
        - pos_profile_id
        - opening_entry_id
        - cashier_party_id
        - movement_type
        - amount
        - moved_at
        - metadata
      properties:
        company_id:
          type: string
          format: uuid
        pos_profile_id:
          type: string
          format: uuid
        opening_entry_id:
          type: string
          format: uuid
        cashier_party_id:
          type: string
          format: uuid
        movement_type:
          type: string
        amount:
          type: string
        reason:
          nullable: true
          type: string
          maxLength: 300
        moved_at:
          type: string
          format: date-time
        metadata:
          type: object
...
    PosCashMovementList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/PosCashMovement'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []