# src/presentation/http/pos_closing_entry_handler.rs
//! PosClosingEntry REST handlers
//!
//! Generated by metaphor-schema. Do not edit manually.
//!
//! Uses Axum and backbone-core's BackboneCrudHandler for all 12 CRUD endpoints.

use std::collections::HashMap;
use std::sync::Arc;

use axum::Router;
use serde::{Deserialize, Serialize};
use uuid::Uuid;
use chrono::{DateTime, Utc};
use rust_decimal::Decimal;

// Backbone framework imports
use backbone_core::http::BackboneCrudHandler;

// Auth integration (optional)
#[cfg(feature = "auth")]
use backbone_auth::middleware::AuthContext;
#[cfg(feature = "auth")]
use backbone_auth::AuthMiddleware;

// Domain imports
use crate::domain::entity::*;
use crate::application::service::{PosClosingEntryService, ServiceError};

// DTO imports
use crate::presentation::dto::{CreatePosClosingEntryDto, UpdatePosClosingEntryDto, PatchPosClosingEntryDto, PosClosingEntryResponseDto};


/// Application error type
#[derive(Debug, thiserror::Error)]
pub enum PosClosingEntryError {
    #[error("Not found: {0}")]
    NotFound(String),
    #[error("Validation error: {0}")]
    Validation(String),
    #[error("Database error: {0}")]
    Database(String),
    #[error("Internal error: {0}")]
    Internal(String),
}

impl From<ServiceError> for PosClosingEntryError {
    fn from(err: ServiceError) -> Self {
        match err {
            ServiceError::NotFound => Self::NotFound(err.to_string()),
            ServiceError::Validation(ref msg) => Self::Validation(msg.clone()),
            ServiceError::AlreadyExists(ref msg) => Self::Validation(msg.clone()),
            ServiceError::Repository(ref e) => Self::Database(e.to_string()),
            ServiceError::Internal(ref msg) => Self::Internal(msg.clone()),
        }
    }
}

impl axum::response::IntoResponse for PosClosingEntryError {
    fn into_response(self) -> axum::response::Response {
        use axum::http::StatusCode;
        use axum::Json;

        let (status, code) = match &self {
            Self::NotFound(_) => (StatusCode::NOT_FOUND, "POSCLOSINGENTRY_NOT_FOUND"),
            Self::Validation(_) => (StatusCode::BAD_REQUEST, "POSCLOSINGENTRY_VALIDATION_ERROR"),
            Self::Database(_) => (StatusCode::INTERNAL_SERVER_ERROR, "POSCLOSINGENTRY_DATABASE_ERROR"),
            Self::Internal(_) => (StatusCode::INTERNAL_SERVER_ERROR, "POSCLOSINGENTRY_INTERNAL_ERROR"),
        };

        let body = serde_json::json!({
            "success": false,
            "error": code,
            "message": self.to_string(),
        });

        (status, Json(body)).into_response()
    }
}

// =============================================================================
// Route Configuration
// =============================================================================

/// Create Axum router with all 16 Backbone endpoints for PosClosingEntry.
///
/// # Routes
///
/// | Method | Path | Description |
/// |--------|------|-------------|
/// | GET | /pos_closing_entries | List with pagination |
/// | POST | /pos_closing_entries | Create new |
/// | GET | /pos_closing_entries/:id | Get by ID |
/// | PUT | /pos_closing_entries/:id | Full update |
/// | PATCH | /pos_closing_entries/:id | Partial update |
/// | DELETE | /pos_closing_entries/:id | Soft delete |
/// | POST | /pos_closing_entries/bulk | Bulk create |
/// | POST | /pos_closing_entries/upsert | Upsert |
/// | GET | /pos_closing_entries/trash | List deleted |
/// | POST | /pos_closing_entries/:id/restore | Restore |
/// | DELETE | /pos_closing_entries/empty | Empty trash |
/// | GET | /pos_closing_entries/:id/deleted | Get deleted by ID |
/// | DELETE | /pos_closing_entries/trash/:id | Permanent delete from trash |
/// | GET | /pos_closing_entries/count | Count active entities |
/// | GET | /pos_closing_entries/trash/count | Count deleted entities |
///
/// # Example
///
/// ```text
/// let service = Arc::new(PosClosingEntryService::with_repository(repository));
/// let router = create_pos_closing_entry_routes(service);
/// ```
pub fn create_pos_closing_entry_routes(service: Arc<PosClosingEntryService>) -> Router {
    BackboneCrudHandler::<PosClosingEntryService, PosClosingEntry, CreatePosClosingEntryDto, UpdatePosClosingEntryDto, PosClosingEntryResponseDto>::routes(
        service,
        "/pos_closing_entries",
    )
}

/// Create Axum router with only the read (GET) endpoints for PosClosingEntry.
///
/// Safe for public, unauthenticated exposure (e.g., reference data).
/// Mutations must be served separately via `create_pos_closing_entry_write_routes`,
/// typically wrapped in an auth middleware layer.
pub fn create_pos_closing_entry_read_routes(service: Arc<PosClosingEntryService>) -> Router {
    BackboneCrudHandler::<PosClosingEntryService, PosClosingEntry, CreatePosClosingEntryDto, UpdatePosClosingEntryDto, PosClosingEntryResponseDto>::read_routes(
        service,
        "/pos_closing_entries",
    )
}

/// Create Axum router with only the write (mutation) endpoints for PosClosingEntry.
///
/// These routes must NOT be publicly exposed. Wrap them with an auth
/// middleware before nesting into the application router.
pub fn create_pos_closing_entry_write_routes(service: Arc<PosClosingEntryService>) -> Router {
    BackboneCrudHandler::<PosClosingEntryService, PosClosingEntry, CreatePosClosingEntryDto, UpdatePosClosingEntryDto, PosClosingEntryResponseDto>::write_routes(
        service,
        "/pos_closing_entries",
    )
}

/// Create authenticated routes with auth middleware.
///
/// Requires the `auth` feature flag. The `AuthMiddleware` implementation
/// is responsible for extracting and validating tokens, then providing
/// an `AuthContext` via request extensions.
#[cfg(feature = "auth")]
pub fn create_protected_pos_closing_entry_routes<A: AuthMiddleware + Send + Sync + 'static>(
    service: Arc<PosClosingEntryService>,
    auth: Arc<A>,
) -> Router {
    use axum::middleware;
    use axum::response::IntoResponse;

    let auth_layer = auth.clone();
    create_pos_closing_entry_routes(service)
        .layer(middleware::from_fn(move |mut req: axum::extract::Request, next: axum::middleware::Next| {
            let auth = auth_layer.clone();
            async move {
                let token = req.headers()
                    .get(axum::http::header::AUTHORIZATION)
                    .and_then(|h| h.to_str().ok())
                    .and_then(|raw| raw.strip_prefix("Bearer ").or_else(|| raw.strip_prefix("bearer ")))
                    .unwrap_or("");
                match auth.authenticate(token).await {
                    Ok(ctx) => {
                        req.extensions_mut().insert(ctx);
                        next.run(req).await
                    }
                    Err(_) => {
                        (axum::http::StatusCode::UNAUTHORIZED,
                         axum::Json(serde_json::json!({
                             "success": false,
                             "error": "unauthorized",
                             "message": "Authentication required"
                         }))
                        ).into_response()
                    }
                }
            }
        }))
}